Create vhosfuzzer.sh
This Bash script is designed to perform Virtual Host (VHOST) fuzzing on a target web server. It iterates over a list of potential VHOST names and sends HTTP requests to the server, checking the response to identify valid virtual hosts. The script utilizes curl to send requests and outputs results in a color-coded format for better readability.